Finished the reorganization process, the project from now on uses a half-realm, half-feature based directory structure
