+++ src/js/server/admin/admin-dashboard.js
@@ -1,9 +1,8 @@
 import m from 'mithril';
 
-import AdminRealm from './admin-realm';
-import AdminMenu from './admin-menu';
+import AdminLayout from './layout/admin-layout';
 
-const AdminDashboard = Object.create(AdminRealm);
+const AdminDashboard = Object.create(AdminLayout);
 
 AdminDashboard.controller = function controller() {
   if (AdminDashboard.checkRedirect()) {
@@ -12,7 +11,7 @@ AdminDashboard.controller = function controller() {
 };
 
 AdminDashboard.view = function view() {
-  return this.constructView([m('h1', 'Hello!'), AdminMenu]);
+  return this.constructView([m('h1', 'Hello!')]);
 };
 
 export default AdminDashboard;
